requirements.txt
openai>=1.40.0
requests>=2.26.0
python-dotenv>=1.0.1
pandas>=1.3.0
openpyxl>=3.1.0